backend-admin/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
pymongo==4.6.0
redis==5.0.1
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
cryptography==41.0.7
pyjwt==2.8.0
django-environ==0.11.2
drf-yasg==1.21.7
markdown==3.5.1
